pyproject.toml
[project]
name = "adktry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.9.0",
    "litellm>=1.74.15.post1",
    "python-dotenv>=1.1.1",
]
